refactor : 관련된 클래스끼리 패키지로 묶음
