+++ src/main/java/blackjack/domain/card/Cards.java
@@ -1,7 +1,9 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
 import java.util.ArrayList;
 import java.util.List;
+
+import blackjack.domain.player.Player;
 
 public class Cards {
