+++ src/test/java/blackjack/domain/CardsTest.java
@@ -9,6 +9,11 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
+import blackjack.domain.card.Card;
+import blackjack.domain.card.CardNumber;
+import blackjack.domain.card.CardPattern;
+import blackjack.domain.card.Cards;
+
 public class CardsTest {
 
     @Nested
