+++ src/test/java/blackjack/domain/NameTest.java
@@ -5,6 +5,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
+import blackjack.domain.player.Name;
+
 public class NameTest {
 
     @Nested
